backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
datasets==2.15.0
peft==0.7.1
bitsandbytes==0.41.3
google-cloud-bigquery==3.13.0
db-dtypes==1.2.0
pyarrow==14.0.1
fastparquet==2023.10.1
duckdb==0.9.2
Pillow==10.1.0
aiohttp==3.9.1
websockets==12.0
